+++ tests/test_conformer_search_data.py
@@ -14,22 +14,26 @@ def test_energies_size(water):
 
     # No energies is fine
     ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         conformers=[w1, w2],
     )
 
     # No energies is fine
     ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         rotamers=[w1, w2],
     )
 
     with pytest.raises(ValidationError):
         ConformerSearchData(
+            provenance={"program": "qcdata-test-suite"},
             conformers=[w1, w2],
             conformer_energies=[-1.0],
         )
 
     with pytest.raises(ValidationError):
         ConformerSearchData(
+            provenance={"program": "qcdata-test-suite"},
             rotamers=[w1, w2],
             rotamer_energies=[-1.0, -2.0, -3.0],
         )
@@ -42,6 +46,7 @@ def test_relative_energies(water):
 
     # No energies returns empty array
     csr = ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         conformers=[w1, w2, w3],
     )
 
@@ -49,6 +54,7 @@ def test_relative_energies(water):
 
     # Energies are relative to the lowest energy
     csr = ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         conformers=[w1, w2, w3],
         conformer_energies=[-1.0, -2.0, -3.0],
     )
@@ -58,6 +64,7 @@ def test_relative_energies(water):
 
     # No energies returns empty array
     csr = ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         rotamers=[w1, w2, w3],
     )
 
@@ -65,6 +72,7 @@ def test_relative_energies(water):
 
     # Energies are relative to the lowest energy
     csr = ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         rotamers=[w1, w2, w3],
         rotamer_energies=[-1.0, -2.0, -3.0],
     )
@@ -79,6 +87,7 @@ def test_conformer_search_results_sorting(water):
     w3 = copy.deepcopy(water)
 
     csr = ConformerSearchData(
+        provenance={"program": "qcdata-test-suite"},
         conformers=[w1, w2, w3],
         conformer_energies=[-1.0, -2.0, -3.0],
     )
@@ -87,4 +96,3 @@ def test_conformer_search_results_sorting(water):
     assert csr.conformers[1] is w2
     assert csr.conformers[2] is w1
     assert csr.conformer_energies.tolist() == [-3.0, -2.0, -1.0]
-